Added additional separation logic for 2, 3 and 4 overlapping characters
In this code i have added speeration logic for overlapping characters from 2 - 4 characters. I do this by segmenting the image into various 'windows' which we can then find the area where the lowest number of black pixels are (we assume this is the best split point). We then store each of these individual characters.

Finally for each section i have created a visualisation so we can see where each of the window boundaries are (blue lines), where the predicted split is (middle of blue lines (yellow lines) ) and the actual split point (green line)